split mnemonic_shift.hpp
